% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_crwidth.R
\name{calc_crwidth}
\alias{calc_crwidth}
\title{Predict tree crown widths using species-specific equations}
\usage{
calc_crwidth(tree_list, digits = 1)
}
\arguments{
\item{tree_list}{A data frame containing tree records. Must have columns
\code{SPCD} (FIA integer species code), \code{STATUSCD} (FIA integer tree status code,
1 = live) and \code{DIA} (FIA tree diameter in inches).}

\item{digits}{Optional integer indicating the number of digits to keep in the
return values (defaults to \code{1}).}
}
\value{
A numeric vector of length \code{nrow(tree_list)} with predicted crown width in
feet for live trees. \code{NA} is returned for trees with \code{STATUSCD != 1}.
}
\description{
\code{calc_crwidth()} computes tree crown width using species-specific
regression equations from the literature.
}
\details{
Crown width is predicted from tree diameter using coefficients provided in
the lookup table \link{cw_coef} (see \code{?cw_coef}) The method also incorporates
adjustment factors used to derive crown width estimates for FIA "saplings",
i.e., trees less than 5.0 in. (12.7 cm) diameter but greater than or equal
to 1.0 in. (2.54 cm) diameter. Details are described in the documentation
for the lookup table \link{cw_sapling_adj}.

Large diameter trees in the temperate rain forests of the Pacific Northwest
region can far exceed the range of diameters in the broadly applicable
datasets that have been used to develop crown width prediction equations
(Bechtold 2003, 2004). To avoid extrapolation beyond the range of the model
fitting data in those cases, \code{calc_crwidth()} makes use of the "old growth"
equation presented by Gill et al. (2000) to estimate crown width for nine
tree species when their diameter is greater than 50 in. (127 cm).
}
\examples{
calc_crwidth(plantation)
}
\references{
Bechtold, W.A. 2003. Crown-diameter prediction models for 87 species of
stand-grown trees in the eastern United States. \emph{Southern Journal of Applied
Forestry}, 27(4): 269-278.

Bechtold, W.A. 2004. Largest-crown-width prediction models for 53 species in
the western United States. \emph{Western Journal of Applied Forestry}, 19(4):
245-251.

Gill, S.J., G.S. Biging, E.C. Murphy. 2000. Modeling conifer tree crown
radius and estimating canopy cover. \emph{Forest Ecology and Management}, 126(3):
405-416.
}
\seealso{
\link{cw_coef}, \link{cw_sapling_adj}
}
