% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tcc_metrics.R
\name{calc_tcc_metrics}
\alias{calc_tcc_metrics}
\title{Predict plot-level canopy cover from individual tree measurements}
\usage{
calc_tcc_metrics(
  tree_list,
  stem_map = TRUE,
  full_output = TRUE,
  digits = 1,
  ...
)
}
\arguments{
\item{tree_list}{A data frame with tree records for one FIA plot. In general,
the input data frame will have the columns specified in
\link{DEFAULT_TREE_COLUMNS} (see \code{?DEFAULT_TREE_COLUMNS}). Potentially, only a
subset of those columns will be needed depending on values given for the
arguments \code{stem_map} and \code{full_output} described below. If the input data
frame has a column named \code{"CRWIDTH"} it will be used for tree crown width
values, otherwise, crown widths will be calculated with a call to
\code{calc_crwidth()}.}

\item{stem_map}{A logical value indicating whether to map individual tree
stems explicitly, using coordinates specified in terms of distance and
azimuth from subplot/microplot centers. The default is \code{TRUE}, in which case
the input \code{tree_list} must contain columns \code{"DIST"} and \code{"AZIMUTH"}. This
argument may be set to \code{FALSE} if individual tree locations are not
available, in which case TCC will be predicted assuming a random arrangement
of tree locations (see Details).}

\item{full_output}{A logical value indicating whether to include the full set
of components used to derive the plot-level prediction. By default, the
output list includes subplot-level TCC estimates, live tree and sapling
counts, stand height metrics, and point pattern statistics, depending on the
value given for \code{stem_map} (see Details).}

\item{digits}{Optional integer indicating the number of digits to keep in the
return values (defaults to \code{1}). May be passed to \code{calc_crwidth()} and
\code{calc_ht_metrics()}.}

\item{...}{Optional arguments passed to \code{create_fia_ppp()} if
\code{stem_map = TRUE}.}
}
\value{
If \code{full_output = TRUE}, a named list with element \code{model_tcc} containing
the plot-level predicted tree canopy cover as percent (\code{0:100}), and
additional named elements containing stand structure metrics as described in
Details. If \code{full_output = FALSE}, a single numeric value of plot-level
predicted TCC is returned instead.
}
\description{
\code{calc_tcc_metrics()} computes predicted plot-level tree canopy cover (TCC)
from standard field inventory measurements. By default, the output includes
a full set of stand structure variables used to model the plot-level TCC
value (see Details).
}
\details{
This function provides two methods for predicting plot-level TCC.

The default "stem-map" method requires individual tree coordinates to be
given in the input as distance and azimuth from subplot centers for trees
with diameter \verb{>= 5 in.} (12.7 cm), and from microplot centers for
"saplings" having diameter \verb{>= 1 in.} (2.54 cm) but \verb{< 5 in.}
(12.7 cm). This method involves mapping trees spatially within the plot
boundary to account for crown overlap explicitly, along with empirical
modeling of the understory sapling contribution to total canopy cover
(Toney et al. 2009). The empirical model for the sapling component also uses
the spatial point pattern of overstory trees as a predictor variable (using
a square root transformation of Ripley's edge-corrected K-function, Ripley
1977, Stoyan and Penttinen 2000).

Alternatively, TCC can be predicted using a simplified approach that does not
include exact stem placement within the plot boundary (\code{stem_map = FALSE}). A
random arrangement of stems is assumed in that case. This is the method used
to estimate tree canopy cover in the Forest Vegetation Simulator (Crookston
and Stage 1999).

Both methods require estimates of individual tree crown widths, which are
computed with \code{calc_crwidth()} if not provided in the input tree list.

The stem-map method also requires computation of several stand structure
metrics which are used in various components of the overall model used to
derive a plot-level TCC estimate. These additional variables include:
\itemize{
\item individual subplot and microplot crown overlays via \code{calc_crown_overlay()}
\item a stand height metric (\code{meanTreeHtBAW}), and plot-level counts of mature
trees and saplings, via \code{calc_ht_metrics()}
\item descriptive spatial statistics for the overstory tree point pattern via
\code{create_fia_ppp() |> spatstat.explore::Lest()}
}

By default, \code{calc_tcc_metrics()} returns a named list containing the
plot-level modeled TCC value, along with those additional component
variables. Specific elements of the returned list include some or all of the
following, conditionally:
\itemize{
\item \verb{$model_tcc}: plot-level predicted canopy cover of trees \verb{>= 1} inch
(\code{2.54} cm)  diameter, derived by one of the two methods described above
depending on the value given for argument \code{stem_map = TRUE|FALSE}
}

If the stem-map method is used, then TCC values derived by crown overlay
on the individual subplot and microplot boundaries are included, along with
means of the four subplot/microplot values:
\itemize{
\item \verb{$subpN_crown_overlay}: estimated canopy cover of trees \verb{>= 5-in.}
(12.7 cm) diameter in subplot \code{N} based on crown overlay (\code{N = 1:4})
\item \verb{$subp_overlay_mean}: mean of the four subplot crown overlays
\item \verb{$micrN_crown_overlay}: estimated canopy cover of saplings in the microplot
of subplot \code{N} based on crown overlay (\code{N = 1:4})
\item \verb{$micr_overlay_mean}: mean of the four microplot crown overlays
}

A set of spatial point pattern statistics is also included when the stem-map
method is used. A square root transformation of Ripley's K function using
isotropic edge correction is computed with \code{spatstat.explore::Lest()} for
trees \verb{>= 5-in.} (12.7 cm) diameter within the four-subplot observation
window. The mean of the following values is a predictor variable in a linear
regression model used to estimate the sapling contribution to total tree
canopy cover:
\itemize{
\item \verb{$L_rft}: estimates of the L-function at \code{r} feet (\code{r} = \code{6}, \code{8}, \code{10},
and \code{12})
}

If the argument \code{full_output = TRUE} (the default), then the output will
also include all of the the named elements from the output of
\code{calc_ht_metrics()}.
}
\examples{
# using the spatially explicit "stem-map model" by default
calc_tcc_metrics(plantation)

# return only the predicted TCC value (`$model_tcc`)
calc_tcc_metrics(plantation, full_output = FALSE)

# using the "FVS method" which assumes random tree locations
calc_tcc_metrics(plantation, stem_map = FALSE, full_output = FALSE)
}
\references{
Crookston, N.L. and A.R. Stage. (1999). Percent canopy cover and stand
structure statistics from the Forest Vegetation Simulator. Gen. Tech. Rep.
RMRS-GTR-24. Ogden, UT: U. S. Department of Agriculture, Forest Service,
Rocky Mountain Research Station. 11 p.
\url{https://research.fs.usda.gov/treesearch/6261}.

Ripley, B.D. (1977). Modelling spatial patterns. \emph{Journal of the Royal
Statistical Society: Series B (Methodological)}, 39(2): 172–192.
\doi{10.1111/j.2517-6161.1977.tb01615.x}.

Stoyan, D., and Penttinen, A. (2000). Recent applications of point process
methods in forestry statistics. \emph{Statistical Science}, 15(1), 61–78.
\url{http://www.jstor.org/stable/2676677}.

Toney, C., J.D. Shaw and M.D. Nelson. 2009. A stem-map model for predicting
tree canopy cover of Forest Inventory and Analysis (FIA) plots. In:
McWilliams, Will; Moisen, Gretchen; Czaplewski, Ray, comps. \emph{Forest Inventory
and Analysis (FIA) Symposium 2008}; October 21-23, 2008; Park City, UT. Proc.
RMRS-P-56CD. Fort Collins, CO: U.S. Department of Agriculture, Forest
Service, Rocky Mountain Research Station. 19 p.
\url{https://research.fs.usda.gov/treesearch/33381}.
}
\seealso{
\code{\link[=calc_crwidth]{calc_crwidth()}}, \code{\link[=calc_crown_overlay]{calc_crown_overlay()}}, \code{\link[=calc_ht_metrics]{calc_ht_metrics()}},
\code{\link[=create_fia_ppp]{create_fia_ppp()}}
}
