% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\alias{ft_to_m}
\alias{m_to_ft}
\alias{in_to_cm}
\alias{cm_to_in}
\title{Convenience functions for unit conversion}
\usage{
ft_to_m(x)

m_to_ft(x)

in_to_cm(x)

cm_to_in(x)
}
\arguments{
\item{x}{Numeric vector of values to convert.}
}
\value{
A numeric vector of the converted values.
}
\description{
Functions to convert between SI and US customary units.
}
\details{
\code{ft_to_m(x)} converts feet to meters.

\code{m_to_ft(x)} converts meters to feet.

\code{in_to_cm(x)} converts inches to centimeters.

\code{cm_to_in(x)} converts centimeters to inches.
}
\examples{
ft_to_m(1)

m_to_ft(1)

in_to_cm(1)

cm_to_in(1)
}
