% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crwidth_coef}
\alias{crwidth_coef}
\title{Regression coefficients for predicting tree crown width}
\format{
\subsection{\code{crwidth_coef}}{

A data frame with 430 rows and 8 columns:
\describe{
\item{symbol}{USDA PLANTS Database species symbol}
\item{SPCD}{FIA tree species code}
\item{common_name}{FIA tree species common name}
\item{surrogate}{Common name of surrogate species if applicable}
\item{b0}{Regression b0 coefficient}
\item{b1}{Regression b1 coefficient}
\item{b1}{Regression b2 coefficient}
\item{reference}{Reference for source of the coefficients}
}
}
}
\usage{
crwidth_coef
}
\description{
A curated set of linear regression coefficients for predicting crown width
from stem diamater of tree species in the conterminous US.
}
\details{
The regression equation is of the general form:

\if{html}{\out{<div class="sourceCode">}}\preformatted{CRWIDTH = b0 + b1 * DIA + b2 * DIA^2
}\if{html}{\out{</div>}}

but with a quadratic term not included in all cases. CRWIDTH is the
predicted tree crown diamater in feet, and DIA is FIA stem diameter in
inches.

In cases that a species-specific equation was not available in the
literature, an equation was assigned from a surrogate species based on
subjectively similar tree physiognomy.
}
\references{
Bechtold, W.A. 2003. Crown-Diameter Prediction Models for 87 Species of
Stand-Grown Trees in the Eastern United States. \emph{Southern Journal of Applied
Forestry}, Vol. 27, No. 4, 269-278.

Bechtold, W.A. 2004. Largest-Crown-Width Prediction Models for 53 Species in
the Western United States. \emph{Western Journal of Applied Forestry}, Vol. 19,
No. 4, 245-251.

Gill, S.J., G.S. Biging, E.C. Murphy. 2000. Modeling conifer tree crown
radius and estimating canopy cover. \emph{Forest Ecology and Management}, 126,
405-416.
}
\keyword{datasets}
