% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cw_sapling_adj}
\alias{cw_sapling_adj}
\title{Sapling crown width adjustment factors}
\format{
\subsection{\code{cw_sapling_adj}}{

A data frame with 23 rows and 5 columns:
\describe{
\item{\code{SPCD}}{\code{\link{integer}}, FIA tree species code}
\item{\code{adj_1inch}}{\code{\link{numeric}},
CW adjustment factor at 1 in. DIA relative to 5 in. DIA}
\item{\code{adj_2inch}}{\code{\link{numeric}},
CW adjustment factor at 2 in. DIA relative to 5 in. DIA}
\item{\code{adj_3inch}}{\code{\link{numeric}},
CW adjustment factor at 3 in. DIA relative to 5 in. DIA}
\item{\code{adj_4inch}}{\code{\link{numeric}},
CW adjustment factor at 4 in. DIA relative to 5 in. DIA}
}
}
}
\source{
Toney et al. 2009. A stem-map model for predicting tree canopy cover of
Forest Inventory and Analysis (FIA) plots.
\url{https://research.fs.usda.gov/treesearch/33381}.
}
\usage{
cw_sapling_adj
}
\description{
A species-specific lookup table of estimated adjustment factors for crown
width (CW) of saplings based on data from Bragg (2001).
}
\details{
FIA "saplings" are trees less than 5.0 in. (12.7 cm) diameter but greater
than or equal to 1.0 in. (2.54 cm) diameter. In general, the data available
to fit regression models predicting crown width (e.g., Bechtold 2003, 2004,
see \link{cw_coef}) do not include trees with diameter less than 5.0 in.
(12.7 cm). Extrapolating beyond the range of the model fitting data is
undesirable, especially since a quadratic term is used in the regression
equations for some species.

Adjustment is based on the proportion of crown width predicted for 5-in.
(12.7 cm) diameter, at each 1-in. (2.54 cm) increment below that.
Intermediate values are interpolated in the crown width prediction method.
Mean adjustment factors are used if a species-specific adjustment is not
available.
}
\examples{
cw_coef[cw_coef$SPCD == 261, ]

cw_sapling_adj[cw_sapling_adj$SPCD == 261, ]
}
\references{
Bragg, D.C. 2001. A local basal area adjustment for crown width prediction.
\emph{Northern Journal of Applied Forestry} 18(1):22-28.
}
\keyword{datasets}
