% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{DEFAULT_TREE_COLUMNS}
\alias{DEFAULT_TREE_COLUMNS}
\title{Names of the FIADB TREE columns used by default in \pkg{FIAstemmap}}
\source{
\url{https://research.fs.usda.gov/products/dataandtools/fia-datamart}
}
\usage{
DEFAULT_TREE_COLUMNS
}
\value{
A character vector of 12 strings containing the names of columns in the
FIADB TREE table used by default in \pkg{FIAstemmap}.
}
\description{
\code{DEFAULT_TREE_COLUMNS} is a character vector of column names in the FIADB
TREE table that are used to compute various tree- and plot-level derived
variables in \pkg{FIAstemmap}. Note that certain outputs can be
generated without this full set of TREE attributes, but these are the ones
generally needed for all functionality in the package.
}
\examples{
DEFAULT_TREE_COLUMNS
}
