% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tree_data.R
\name{load_tree_data}
\alias{load_tree_data}
\title{Load tree data from a file or database connection}
\usage{
load_tree_data(
  src,
  table = NULL,
  columns = DEFAULT_TREE_COLUMNS,
  sql = NULL,
  quoted_cols_as_char = TRUE,
  colnames_toupper = TRUE
)
}
\arguments{
\item{src}{A character string specifying the data source as a file name or
database connection string (see Details).}

\item{table}{Optional character string giving the name of a table in \code{src}
from which tree records will be fetched. Generally needed with database
sources containing multiple tables (as opposed to a single-table source such
as a CSV file).}

\item{columns}{Optional character vector specifying a subset of column names
in the source table to include in the result set. Defaults to
\link{DEFAULT_TREE_COLUMNS}. Can also be set to \code{NULL} or empty string (\code{""}) to
read all columns in the source table.}

\item{sql}{Optional character string containing a SQL SELECT statement to
execute on \code{src} (instead of selecting all records, potentially from a subset
of columns, i.e., mutually exclusive with \code{table} and/or \code{columns}).}

\item{quoted_cols_as_char}{A logical value indicating whether to auto-detect
columns that contain quoted values as \code{"character"} type, \code{TRUE} by default.
Only used when \code{src} is a CSV file.}

\item{colnames_toupper}{Logical value, \code{TRUE} to enforce upper case column
names in the returned data frame (the default).}
}
\value{
A data frame containing tree records fetched from \code{src}.
}
\description{
\code{load_tree_data()} fetches tree records from a data source, typically a
comma-separated values (CSV) file, a SQLite database file (.db, .sqlite,
.gpkg), or a PostgreSQL database connection. Other data sources are also
possible. File-based sources can be read from compressed archives without
prior extraction if desired (e.g., .zip), and network-hosted files can be
read directly without prior download (see Details).
}
\details{
A data source is most commonly specified as one of the following:

CSV file\cr
Path to a text file with \code{".csv"} extension. For files structured as CSV,
but not ending with the \code{".csv"} extension, a \code{"CSV:"} prefix can be added
before the filename to force loading as CSV format.

SQLite database\cr
Path to a SQLite file. The file extension is generally either \code{".db"} or
\code{".sqlite"}, but GeoPackage files with the \code{".gpkg"} extension are also
supported.

PostgreSQL database\cr
A connection string in one of the following formats:

\if{html}{\out{<div class="sourceCode">}}\preformatted{"PG:dbname=databasename"

"PG:dbname='db' host='addr' port='5432' user='x' password='y'"

"PG:service=servicename"

"postgresql://[user[:pwd]@][netloc][:port][/dbname][?param1=val1&...]"
}\if{html}{\out{</div>}}

GDAL Virtual File Systems are also supported. This allows, for example,
reading from compressed archives such as \code{".zip"} without prior extraction.
The syntax in that case uses the \code{"/vsizip/"} prefix:

\if{html}{\out{<div class="sourceCode">}}\preformatted{# relative path to the .zip:
f <- "/vsizip/MT_CSV.zip/MT_TREE.csv"

# absolute path to the .zip:
f <- "/vsizip//home/ctoney/data/MT_CSV.zip/MT_TREE.csv"

# on Windows:
f <- "/vsizip/c:/users/ctoney/MT_CSV.zip/MT_TREE.csv"
}\if{html}{\out{</div>}}

Network-hosted files can also be read without prior download using the
\code{"/vsicurl/"} prefix:

\if{html}{\out{<div class="sourceCode">}}\preformatted{f <- "/vsicurl/https://apps.fs.usda.gov/fia/datamart/CSV/MT_TREE.csv"
}\if{html}{\out{</div>}}

For more details, including supported VSI prefixes for cloud storage services
and other virtual file systems, see
\url{https://gdal.org/en/stable/user/virtual_file_systems.html}.
}
\note{
\code{src} can be any GDAL supported dataset. A full list of formats supported by
the current GDAL installation can be obtained with:

\if{html}{\out{<div class="sourceCode">}}\preformatted{fmt <- gdalraster::gdal_formats()
fmt$long_name[fmt$vector]
}\if{html}{\out{</div>}}

For more details: \url{https://gdal.org/en/stable/drivers/vector/index.html}

\code{load_tree_data()} from a PostgreSQL database requires GDAL built with
support for the PostgreSQL client library (can be checked with
\code{gdalraster::gdal_formats("postgresql")}).

Column names are case-sensitive in \pkg{FIAstemmap} functions, and are
assumed to follow the FIADB upper case naming convention.

If column \code{PLT_CN} is present, it will be read or coerced if necessary to R
\code{"character"} type consistent with its data type in FIADB (i.e., a string
but all digits).
}
\examples{
# Lolo NF, single-condition forest plots, INVYR 2022, from public FIADB
f <- system.file("extdata/mt_lnf_2022_1cond_tree.csv", package="FIAstemmap")
tree <- load_tree_data(f)

head(tree)
}
\seealso{
\link{DEFAULT_TREE_COLUMNS},\code{\link[=process_tree_data]{process_tree_data()}}
}
