% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tree_data.R
\name{process_tree_data}
\alias{process_tree_data}
\title{Generate plot-level stand structure metrics for a tree dataset}
\usage{
process_tree_data(tree_table, stem_map = TRUE, full_output = TRUE, digits = 1)
}
\arguments{
\item{tree_table}{A data frame containing tree records for a set of forest
inventory plots. Must have column \code{PLT_CN} containing the plot unique
identifier for each tree. Other required columns are those of
\code{calc_crwidth()} (if column \code{CRWIDTH} is not included), \code{calc_ht_metrics()}
and \code{calc_tcc_metrics()}, depending on values given for \code{stem_map} and
\code{full_output}.}

\item{stem_map}{A logical value indicating whether to map individual tree
stems explicitly, using coordinates specified in terms of distance and
azimuth from subplot/microplot centers. The default is \code{TRUE}, in which case
the input \code{tree_table} must contain columns \code{"DIST"} and \code{"AZIMUTH"}. This
argument may be set to \code{FALSE} if individual tree locations are not
available, in which case TCC will be predicted assuming a random arrangement
of tree locations (see Details for \code{\link[=calc_tcc_metrics]{calc_tcc_metrics()}}).}

\item{full_output}{A logical value indicating whether to include the full set
of components used to derive the plot-level TCC prediction. By default, the
output data includes subplot-level TCC estimates, live tree and sapling
counts, stand height metrics, and point pattern statistics, depending on the
value given for \code{stem_map} (see Details for \code{\link[=calc_tcc_metrics]{calc_tcc_metrics()}}).}

\item{digits}{Optional integer indicating the number of digits to keep in the
return values (defaults to \code{1}). May be passed to \code{calc_crwidth()} and
\code{calc_ht_metrics()}.}
}
\value{
A data frame with one row for each unique \code{PLT_CN} in the input \code{tree_table},
and additional columns containing the output of \code{calc_tcc_metrics()}
conditional on the values given for \code{stem_map} and \code{full_output}.
}
\description{
\code{process_tree_data()} takes a table of tree records for a set of forest
inventory plots as input, and generates selected plot-level stand structure
metrics.
}
\examples{
# Lolo NF, single-condition forest plots, INVYR 2022, from public FIADB
f <- system.file("extdata/mt_lnf_2022_1cond_tree.csv", package="FIAstemmap")
tree_table <- load_tree_data(f)

process_tree_data(tree_table, stem_map = FALSE, full_output = TRUE)
}
\seealso{
\code{\link[=calc_ht_metrics]{calc_ht_metrics()}}, \code{\link[=calc_tcc_metrics]{calc_tcc_metrics()}}, \code{\link[=load_tree_data]{load_tree_data()}}
}
