% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{western_redcedar}
\alias{western_redcedar}
\title{Tree list for a western redcedar forest}
\format{
\subsection{\code{western_redcedar}}{

A data frame with 33 rows and 12 columns:
\describe{
\item{\code{PLT_CN}}{\code{\link{character}}, Plot unique identifier}
\item{\code{SUBP}}{\code{\link{integer}}, Subplot number}
\item{\code{TREE}}{\code{\link{integer}}, Tree number}
\item{\code{AZIMUTH}}{\code{\link{integer}},
Horizontal angle from subplot center to the stem location}
\item{\code{DIST}}{\code{\link{numeric}},
Distance in feet from subplot center to the stem location}
\item{\code{STATUSCD}}{\code{\link{integer}},
Tree status code: 1 = live, 2 = standing dead}
\item{\code{SPCD}}{\code{\link{integer}}, FIA tree species code}
\item{\code{DIA}}{\code{\link{numeric}},
Tree diameter at breast height in inches}
\item{\code{HT}}{\code{\link{numeric}}, Tree height in feet}
\item{\code{ACTUALHT}}{\code{\link{numeric}},
Actual height in ft (ACTUALHT < HT indicates a broken top)}
\item{\code{CCLCD}}{\code{\link{integer}}, Tree crown class code}
\item{\code{TPA_UNADJ}}{\code{\link{numeric}},
Trees per acre expansion factor}
}
}
}
\source{
\url{https://research.fs.usda.gov/programs/nfi}
}
\usage{
western_redcedar
}
\description{
An example tree list for an FIA plot in a western redcedar forest.
}
\note{
A synthetic plot unique identifier is used in example tree list datasets.
}
\keyword{datasets}
