% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cbh_metrics.R
\name{calc_cbh_metrics}
\alias{calc_cbh_metrics}
\title{Calculate canopy base height metrics from FIA tree data}
\usage{
calc_cbh_metrics(tree_table, plot_id_col = NULL, digits = 1)
}
\arguments{
\item{tree_table}{A data frame with tree records for one or more FIA plots.
Must have columns \code{ACTUALHT} (tree actual height) and \code{UNCRCD} (uncompacted
live crown ratio as percentage).}

\item{plot_id_col}{Optional column name in \code{tree_table} of a plot unique
identifier. If \code{NULL} and no column named \code{PLT_CN} or \code{plt_cn} exists,
the input data are assumed to be a single tree list. If \code{PLT_CN} or \code{plt_cn}
exists, it will be used as the plot ID column.}

\item{digits}{Optional integer indicating the number of digits to keep in the
return values (defaults to \code{1}).}
}
\value{
Canopy base height metrics aggregated from individual tree crown base
heights: \code{"cbh_mean"}, \code{"cbh_median"}, \code{"cbh_pct20"} and \code{"cbh_min"}. If the
input contains tree data for a single plot without a plot ID column, a named
list is returned, otherwise a data frame.
}
\description{
\code{calc_cbh_metrics()} computes canopy base height metrics for FIA tree data
based on measurements of uncompacted live crown ratio (\code{UNCRCD}) of
individual trees \verb{>= 5.0} in. (\code{12.7} cm) diameter. Canopy base height
represents the lowest height at which canopy fuel, aggregated across multiple
tree crowns, can propagate fire vertically (see, e.g., Mast et al. 2026).
This function derives individual tree crown base heights from actual tree
height and the uncompacted crown ratio, and computes the mean, median,
\eqn{20^{th}}{20th} percentile and minimum.
}
\examples{
f <- system.file("extdata/dfir_plot.csv", package="FIAstemmap")
trees <- load_tree_data(f, columns = NULL)
calc_cbh_metrics(trees)
}
\references{
Mast, C.N., L. Vallet, M.D. Hurteau, H. Zald, G.M. Jones, G.A. Woolsey,
W. Tinkham, C.M. Hoffman, E.J. Francis. 2026. Methods and implications of
canopy base height characterization for modeling crown fire potential.
\emph{Fire Ecology} 22(117). \doi{10.1186/s42408-026-00567-4}.
}
\seealso{
\code{\link[=calc_ht_metrics]{calc_ht_metrics()}}, \code{\link[=calc_tcc_metrics]{calc_tcc_metrics()}}
}
