% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ht_metrics.R
\name{calc_ht_metrics}
\alias{calc_ht_metrics}
\title{Calculate stand height metrics}
\usage{
calc_ht_metrics(tree_list, digits = 1)
}
\arguments{
\item{tree_list}{A data frame with tree records for one FIA plot.  Must have
columns \code{DIA} (tree diameter), \code{HT} (tree height), \code{ACTUALHT} (tree actual
height, \code{ACTUALHT < HT} indicating a broken top), \code{CCLCD} (FIA crown class
code), \code{TPA_UNADJ} (trees per acre).}

\item{digits}{Optional integer indicating the number of digits to keep in the
return values (defaults to \code{1}).}
}
\value{
A named list of computed height metrics for the input tree list, as described
in Details.
}
\description{
\code{calc_ht_metrics()} computes several stand height metrics for a given tree
list (see Details).
}
\details{
Stand height metrics are based on live trees (\code{STATUSCD == 1}), and are
are assigned \code{0} by definition if no live trees are present. Height metrics
are returned in a named list with the following elements:
\itemize{
\item \code{numTrees}: number of live trees \verb{>= 5.0} in. (\code{12.7} cm) diameter
\item \code{meanTreeHt}: mean height of trees \verb{>= 5.0} in. (\code{12.7} cm) diameter
\item \code{meanTreeHtBAW}: basal-area weighted mean height of trees \verb{>= 5.0} in.
(\code{12.7} cm) diameter
\item \code{meanTreeHtDom}: mean height of canopy dominant/co-dominant trees
\verb{>= 5.0} in. (\code{12.7} cm) diameter
\item \code{meanTreeHtDomBAW}: basal-area weighted mean height of canopy
dominant/co-dominant trees \verb{>= 5.0} in. (\code{12.7} cm) diameter
\item \code{maxTreeHt}: height of the tallest tree \verb{>= 5.0} in. (\code{12.7} cm) diameter
\item \code{predomTreeHt}: predominant tree height, as the mean height of the tallest
trees \verb{>= 5.0} in. (\code{12.7} cm) diameter comprising up to \code{16} trees per
acre (\code{39.5} trees per hectare)
\item \code{numSaplings}: number of live saplings (trees \verb{>= 1.0} in. but \verb{< 5.0} in.
diameter, i.e., \verb{>= 2.54} cm but \verb{< 12.7} cm)
\item \code{meanSapHt}: mean height of saplings
\item \code{maxSapHt}: height of the tallest sapling
}

For the purpose of height calculations, metrics based on
"canopy dominant/co-dominant" include open grown trees, i.e., include trees
with FIA crown class code \code{CCLCD} of \code{1} (open grown), \code{2} (dominant) or \code{3}
(co-dominant), but exclude trees with \code{CCLCD} of \code{4} (intermediate) or \code{5}
(over-topped).
}
\examples{
calc_ht_metrics(plantation)

}
