% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fia_owin.R
\name{create_fia_owin}
\alias{create_fia_owin}
\title{Create a 'spatstat' observation window for the FIA plot footprint}
\usage{
create_fia_owin(linear_unit = "ft", macroplot = FALSE)
}
\arguments{
\item{linear_unit}{An optional character string specifying the linear
distance unit. Defaults to the native FIA unit of \verb{"ft"``, but may be set to }"m"`.}

\item{macroplot}{An optional logical value. The default is \code{FALSE} with the
FIA plot footprint defined in terms of the standard four-subplot
configuration with subplot radius of 24 ft (7.3152 m). May be set to \code{TRUE}
in which case the observation window will be defined in terms of the FIA
optional macroplot dimension instead (58.9 ft / 18.227 m raduis, used only
in certain areas of the Pacific Northwest FIA region).}
}
\value{
An object of class \code{"owin"} which defines the observation window for a point
pattern in the \pkg{spatstat} library.
}
\description{
\code{create_fia_owin()} returns an object of class \code{"owin"} representing the
observation window of an FIA plot in the 2-D plane. Class \code{"owin"} defines
such window for working with an FIA tree list as a point pattern dataset
via functions of the \pkg{spatstat} package.
}
\examples{
w <- create_fia_owin()
plot(w, main = "FIA standard four-subplot design")

}
