% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fia_ppp.R
\name{create_fia_ppp}
\alias{create_fia_ppp}
\title{Create a 'spatstat' point pattern dataset for an FIA plot tree list}
\usage{
create_fia_ppp(
  tree_list,
  live_only = TRUE,
  min_dia = 5,
  linear_unit = "ft",
  macroplot = FALSE,
  window = create_fia_owin(linear_unit, macroplot),
  mark_cols = c("SPCD", "DIA", "ACTUALHT", "CRWIDTH"),
  mark_as_factor = c("SPCD")
)
}
\description{
\code{create_fia_ppp()} returns an object of class \code{"ppp"} representing the point
pattern of an FIA tree list in the 2-D plane. A point pattern object defines
the dataset for using a stem-mapped FIA plot with functions of the
\pkg{spatstat} library.
}
