% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_overlay_pct.R
\name{crown_overlay_pct}
\alias{crown_overlay_pct}
\title{Compute tree canopy cover of a subplot/microplot by crown overlay}
\usage{
crown_overlay_pct(plot_radius, trees)
}
\description{
Compute tree canopy cover of a subplot/microplot by crown overlay
}
