% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crwidth_coef}
\alias{crwidth_coef}
\title{Regression coefficients for predicting tree crown width}
\format{
\subsection{\code{crwidth_coef}}{

A data frame with 430 rows and 8 columns:
\describe{
\item{\code{symbol}}{\code{\link{character}},
USDA PLANTS Database species symbol}
\item{\code{SPCD}}{\code{\link{integer}}, FIA tree species code}
\item{\code{common_name}}{\code{\link{character}},
FIA tree species common name}
\item{\code{surrogate}}{\code{\link{character}},
Common name of surrogate species if applicable}
\item{\code{b0}}{\code{\link{numeric}}, Regression b0 coefficient}
\item{\code{b1}}{\code{\link{numeric}}, Regression b1 coefficient}
\item{\code{b1}}{\code{\link{numeric}}, Regression b2 coefficient}
\item{\code{reference}}{\code{\link{character}},
Literature source of the species coefficients (see References)}
}
}
}
\usage{
crwidth_coef
}
\description{
A curated set of linear regression coefficients for predicting crown width
from stem diameter of tree species in the conterminous US.
}
\details{
The regression equation is of the general form:

\if{html}{\out{<div class="sourceCode">}}\preformatted{CRWIDTH = b0 + b1 * DIA + b2 * DIA^2
}\if{html}{\out{</div>}}

A quadratic term is not included in all cases. \code{CRWIDTH} is the
predicted tree crown diameter in feet, and \code{DIA} is FIA stem diameter in
inches.

In cases that species-specific equations were not available in the
literature, surrogate species were assigned based on subjectively similar
tree physiognomy.
}
\references{
Bechtold, W.A. 2003. Crown-Diameter Prediction Models for 87 Species of
Stand-Grown Trees in the Eastern United States. \emph{Southern Journal of Applied
Forestry}, Vol. 27, No. 4, 269-278.

Bechtold, W.A. 2004. Largest-Crown-Width Prediction Models for 53 Species in
the Western United States. \emph{Western Journal of Applied Forestry}, Vol. 19,
No. 4, 245-251.

Gill, S.J., G.S. Biging, E.C. Murphy. 2000. Modeling conifer tree crown
radius and estimating canopy cover. \emph{Forest Ecology and Management}, 126,
405-416.
}
\keyword{datasets}
