% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tree_data.R
\name{load_tree_data}
\alias{load_tree_data}
\title{Load tree data from a file or database connection}
\usage{
load_tree_data(src, table = NULL, columns = DEFAULT_TREE_COLUMNS, sql = NULL)
}
\arguments{
\item{src}{A character string specifying the data source as a file name or
database connection string (see Details).}

\item{table}{Optional character string giving the name of a table in \code{src}
from which tree records will be fetched. Generally needed with RDBMS sources
containing multiple tables, as opposed to a single-table source such as a
CSV file.}

\item{columns}{Optional character vector specifying a subset of column names
in the source table to include in the result set.}

\item{sql}{Optional character string containing a SQL SELECT statement to
execute on \code{src} (instead of selecting all records potentially from a subset
of columns, i.e., mutually exclusive with \code{table} and/or \code{columns}).}
}
\value{
A data frame containing the tree records fetched from \code{src}.
}
\description{
\code{load_tree_data()} fetches tree records from a data source, most commonly
a comma-separated values (CSV) file, a SQLite database file (.db, .sqlite,
.gpkg), or a PostgreSQL database connection. Other data sources are also
possible. File-based sources can be read from compressed archives without
prior extraction (e.g., .zip), and network-hosted files can be read directly
without prior download (see Details).
}
\details{
A data source is typically specified as one of the following: \strong{WIP}
}
