% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crowns.R
\name{plot_crowns}
\alias{plot_crowns}
\title{Display modeled tree crowns projected vertically on subplot boundaries}
\usage{
plot_crowns(
  tree_list,
  subplot = NULL,
  microplot = FALSE,
  linear_unit = "ft",
  main = "",
  crown_col = "#328e13",
  stem_col = "#b85e00",
  subp_border_lwd = 3,
  subp_border_col = "gray61"
)
}
\arguments{
\item{tree_list}{A data frame with tree records for one FIA plot.  Must have
columns \code{SUBP} (FIA subplot number), \code{STATUSCD} (FIA integer tree status,
\code{1} = live), \code{DIA} (tree diameter), \code{HT} (tree height), \code{ACTUALHT} (tree
actual height, \code{ACTUALHT < HT} indicating a broken top), \code{DIST} (stem
distance from subplot/microplot center) and \code{AZIMUTH} (horizontal angle from
subplot/microplot center to the stem location, in range \code{0:359}). If the
input data frame has a column named \code{"CRWIDTH"} it will be used for tree
crown width values, otherwise, crown widths will be calculated with a call
to \code{calc_crwidth()}.}

\item{subplot}{Optional integer subplot number in the range \code{1:4} indicating
a specific subplot for display. May be \code{NULL} or \code{NA} to display the whole
four-point cluster plot.}

\item{microplot}{A logical value, \code{TRUE} to display the modeled crowns of
saplings overlaid of the microplot boundary of \code{subplot = n}. The default is
\code{FALSE}. Ignored if \code{subplot} is not specified.}

\item{linear_unit}{An optional character string specifying the linear
distance unit. Defaults to the native FIA unit of \code{"ft"}, but may be set to
\code{"m"} instead (or \code{"meter"} / \code{"metre"}), in which case subplot boundaries
will be display in meters, tree heights and crown widths are assumed to be
given in meters, and tree diameters are assumed to be given in centimeters.
\strong{TODO: not currently implemented}}

\item{main}{Character string giving the main plot title (on top).}

\item{crown_col}{The color of tree crowns, e.g., either a color name (as
listed by \code{colors()}) or a hexadecimal string.}

\item{stem_col}{The color of tree stems when plotting an individual subplot
or microplot (see \code{crown_col} above).}

\item{subp_border_lwd}{The line width of subplot boundaries. Must a positive
number.}

\item{subp_border_col}{The color of subplot boundaries (see \code{crown_col}
above).}
}
\value{
The input, invisibly.
}
\description{
\code{plot_crowns()} draws vertically projected tree crowns as discs overlaid on
FIA subplot or microplot boundaries. The full four-subplot cluster, or
individual subplots, can be displayed with trees \verb{>= 5.0} in. (\code{12.7} cm)
diameter. Individual microplots can also be displayed with saplings (i.e.,
trees \verb{< 5} in. diameter).
}
\examples{
plot_crowns(plantation, main = "plantation plot")

plot_crowns(plantation, subplot = 4, main = "plantation subplot 4")

plot_crowns(plantation, subplot = 4, microplot = TRUE,
            main = "plantation microplot 4")
}
\seealso{
\code{\link[=calc_crwidth]{calc_crwidth()}}, \code{\link[=calc_crown_overlay]{calc_crown_overlay()}}
}
