% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatstat_helpers.R
\name{spatstat_helpers}
\alias{spatstat_helpers}
\alias{create_fia_owin}
\alias{create_fia_ppp}
\title{Analyze the spatial pattern of trees on an FIA plot}
\usage{
create_fia_owin(linear_unit = "ft", macroplot = FALSE, npoly = 360)

create_fia_ppp(
  tree_list,
  live_trees = TRUE,
  min_dia = 5,
  linear_unit = "ft",
  macroplot = FALSE,
  window = NULL,
  mark_cols = NULL,
  mark_as_factor = NULL
)
}
\arguments{
\item{linear_unit}{An optional character string specifying the linear
distance unit. Defaults to the native FIA unit of \code{"ft"}, but may be set to
\code{"m"} instead (or \code{"meter"} / \code{"metre"}).}

\item{macroplot}{An optional logical value. The default is \code{FALSE}, which
defines the FIA plot footprint in terms of the standard four-subplot
configuration with subplot radius of 24 ft (7.3152 m). By default,
"macroplot trees" having \code{DIST} outside the subplot boundary are not
included. This argument may be set to \code{TRUE} in which case the observation
window will be defined using the FIA optional "macroplot" configuration
instead (58.9 ft or 18.227 m radius, used only in certain areas of the
Pacific Northwest FIA region).}

\item{npoly}{Integer value giving the number of edges to use for polygon
approximation. Defaults to \code{360}.}

\item{tree_list}{A data frame containing a set of tree records for one FIA
plot (see Details).}

\item{live_trees}{A logical value, \code{TRUE} to include live trees only (the
default, i.e., \code{STATUSCD == 1}).}

\item{min_dia}{A numeric value specifying the minimum diameter threshold
for included trees. The default is \code{5.0}. Trees less than 5-in. diameter but
greater than or equal to 1.0-in. diameter, denoted as "saplings", are only
recorded in FIA microplots so cannot be stem-mapped across the full 4-subplot
footprint.}

\item{window}{An optional object of class \code{"owin"} defining the observation
window of an FIA plot in the 2-D plane. Defaults to
\code{create_fia_owin(linear_unit, macroplot)}.}

\item{mark_cols}{An optional character vector of column names in \code{tree_list}
to designate as \pkg{spatstat} \code{marks} which carry additional information for
each data point in a point pattern object.}

\item{mark_as_factor}{An optional subset of \code{mark_cols} to be treated as
\code{factor} marks. If not already \code{factor}, these will be coerced as such upon
input. \code{factor} marks are those that take only a finite number of possible
values (e.g. colors or types).}
}
\description{
These functions facilitate point pattern analysis of FIA tree data using the
\pkg{spatstat} library.
}
\details{
\code{create_fia_owin()} returns an object of class \code{"owin"} from the
\pkg{spatstat} library. This object represents the generic 2-D observation
window for the nationally standard FIA plot design which is a four-point
cluster of circular subplots. Used when creating a \pkg{spatstat} point
pattern object for an FIA tree list.

\code{create_fia_ppp()} returns an object of class \code{"ppp"} representing the point
pattern of an FIA tree list in the 2-D plane. A point pattern object defines
the dataset for using a stem-mapped FIA plot with functions of the
\pkg{spatstat} library.

The standard set of columns for tree list data in \pkg{FIAstemmap} is given
below, along with the status of each column as required or optional for
\code{create_fia_ppp()}:
\itemize{
\item \code{PLT_CN}: optional, \code{create_fia_ppp()} assumes input is for one plot
\item \code{SUBP}: required subplot number
\item \code{TREE}: required tree number within a subplot
\item \code{AZIMUTH}: required horizontal angle from subplot center
\item \code{DIST}: required distance from subplot center
\item \code{STATUSCD}: required tree status code (1 = live, 2 = dead)
\item \code{SPCD}: optional FIA species code
\item \code{DIA}: optional tree diameter
\item \code{HT}: optional tree height
\item \code{ACTUALHT}: optional tree actual height (accounts for broken top)
\item \code{CCLCD}: optional crown class code
\item \code{TPA_UNADJ}: optional tree expansion factor (per acre)
\item \code{CRWIDTH}: optional crown width (see \code{calc_crwidth()})
}
}
\examples{
## observation window for the standard FIA plot design
w <- create_fia_owin()
summary(w)

# or using metric units
w <- create_fia_owin("m")
summary(w)

plot(w, main = "FIA standard four-subplot design")

## point pattern object for the `plantation` example data
X <- create_fia_ppp(plantation)
summary(X)

plot(X, pch = 16, main = "Loblolly pine plantation")

# plot trees as trees :)
X <- create_fia_ppp(plantation, mark_cols = "SPCD")
plot(X, shape = "arrows", direction = 90, size = 12, cols = "darkgreen",
     legend = FALSE, main = "Loblolly pine plantation")

# Ripley's K-function
K <- spatstat.explore::Kest(X, rmax = 12, correction = "isotropic")
plot(K, main = "Ripley's K for the plantation FIA plot")

## point pattern object for the `western_redcedar` example data
## tree distances may be specified in meters
trees <- within(western_redcedar, DIST <- ft_to_m(DIST))

X <- create_fia_ppp(trees, linear_unit = "m")
summary(X)

plot(X, pch = 16, main = "Western redcedar FIA plot")

# Ripley's K-function
K <- spatstat.explore::Kest(X, rmax = ft_to_m(12), correction = "isotropic")
plot(K, main = "Ripley's K for the western redcedar FIA plot")
}
\references{
Baddeley, A., E. Rubak, R. Turner. 2015. \emph{Spatial Point Patterns: Methodology
and Applications with R}. Chapman and Hall/CRC Press, London.
ISBN 9781482210200,
\url{https://www.routledge.com/Spatial-Point-Patterns-Methodology-and-Applications-with-R/Baddeley-Rubak-Turner/p/book/9781482210200/}.

Ripley, B.D. 1977. Modelling spatial patterns. \emph{Journal of the Royal
Statistical Society: Series B (Methodological)}, 39(2): 172–192.
\url{https://doi.org/10.1111/j.2517-6161.1977.tb01615.x}.
}
\seealso{
\pkg{spatstat} package: \url{https://cran.r-project.org/package=spatstat}
}
